+++ skbuild/resources/cmake/FindNumPy.cmake
@@ -1,20 +1,35 @@
 #.rst:
 #
-# Find the include directory for numpy/arrayobject.h
+# Find the include directory for ``numpy/arrayobject.h`` as well as other NumPy tools like ``conv-template`` and
+# ``from-template``.
 #
 # This module sets the following variables:
 #
 # ``NumPy_FOUND``
 #   True if NumPy was found.
-# ``NumPy_INCLUDE_DIR``
+# ``NumPy_INCLUDE_DIRS``
 #   The include directories needed to use NumpPy.
 # ``NumPy_VERSION``
 #   The version of NumPy found.
-#
+# ``NumPy_CONV_TEMPLATE_EXECUTABLE``
+#   Path to conv-template executable.
+# ``NumPy_FROM_TEMPLATE_EXECUTABLE``
+#   Path to from-template executable.
 #
 # The module will also explicitly define one cache variable:
 #
 # ``NumPy_INCLUDE_DIR``
+#
+# .. note::
+#
+#     To support NumPy < v0.15.0 where ``from-template`` and ``conv-template`` are not declared as entry points,
+#     the module emulates the behavior of standalone executables by setting the corresponding variables with the
+#     path the the python interpreter and the path to the associated script. For example:
+#     ::
+#
+#         set(NumPy_CONV_TEMPLATE_EXECUTABLE /path/to/python /path/to/site-packages/numpy/distutils/conv_template.py CACHE STRING "Command executing conv-template program" FORCE)
+#
+#         set(NumPy_FROM_TEMPLATE_EXECUTABLE /path/to/python /path/to/site-packages/numpy/distutils/from_template.py CACHE STRING "Command executing from-template program" FORCE)
 #
 
 if(NOT NumPy_FOUND)
@@ -27,6 +42,9 @@ if(NOT NumPy_FOUND)
   endif()
   find_package(PythonInterp ${_find_extra_args})
   find_package(PythonLibs ${_find_extra_args})
+
+  find_program(NumPy_CONV_TEMPLATE_EXECUTABLE NAMES conv-template)
+  find_program(NumPy_FROM_TEMPLATE_EXECUTABLE NAMES from-template)
 
   if(PYTHON_EXECUTABLE)
     execute_process(COMMAND "${PYTHON_EXECUTABLE}"
@@ -41,6 +59,28 @@ if(NOT NumPy_FOUND)
       OUTPUT_STRIP_TRAILING_WHITESPACE
       ERROR_QUIET
       )
+
+    # XXX This is required to support NumPy < v0.15.0. See note in module documentation above.
+    if(NOT NumPy_CONV_TEMPLATE_EXECUTABLE)
+      execute_process(COMMAND "${PYTHON_EXECUTABLE}"
+        -c "from numpy.distutils import conv_template; print(conv_template.__file__)"
+        OUTPUT_VARIABLE _numpy_conv_template_file
+        OUTPUT_STRIP_TRAILING_WHITESPACE
+        ERROR_QUIET
+        )
+      set(NumPy_CONV_TEMPLATE_EXECUTABLE "${PYTHON_EXECUTABLE}" "${_numpy_conv_template_file}" CACHE STRING "Command executing conv-template program" FORCE)
+    endif()
+
+    # XXX This is required to support NumPy < v0.15.0. See note in module documentation above.
+    if(NOT NumPy_FROM_TEMPLATE_EXECUTABLE)
+      execute_process(COMMAND "${PYTHON_EXECUTABLE}"
+        -c "from numpy.distutils import from_template; print(from_template.__file__)"
+        OUTPUT_VARIABLE _numpy_from_template_file
+        OUTPUT_STRIP_TRAILING_WHITESPACE
+        ERROR_QUIET
+        )
+      set(NumPy_FROM_TEMPLATE_EXECUTABLE "${PYTHON_EXECUTABLE}" "${_numpy_from_template_file}" CACHE STRING "Command executing from-template program" FORCE)
+    endif()
   endif()
 endif()
 
@@ -56,7 +96,10 @@ set(NumPy_INCLUDE_DIRS ${NumPy_INCLUDE_DIR})
 # all listed variables are TRUE
 include(FindPackageHandleStandardArgs)
 find_package_handle_standard_args(NumPy
-                                  REQUIRED_VARS NumPy_INCLUDE_DIR
+                                  REQUIRED_VARS
+                                    NumPy_INCLUDE_DIR
+                                    NumPy_CONV_TEMPLATE_EXECUTABLE
+                                    NumPy_FROM_TEMPLATE_EXECUTABLE
                                   VERSION_VAR NumPy_VERSION
                                   )